pyproject.toml
[tool.poetry]
name = "md2gost"
version = "0.1.0"
description = ""
authors = ["witelokk <death.thunder@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-docx = "^0.8.11"
marko = "^2.0.0"
docxcompose = "^1.4.0"
freetype-py = "^2.4.0"
pillow = "^10.0.0"
matplotlib = { version = "3.7.2", platform = "!= linux" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tools.poetry.scripts]
md2gost = "md2gost.__main__.main"
